netsvc subagent included into Windows build; 1.2.12 release preparation

+++ src/agent/subagents/netsvc/netsvc.h
@@ -32,8 +32,8 @@ enum
 
 typedef struct
 {
-   INT32 size;
-   INT32 allocated;
+   size_t size;
+   size_t allocated;
    char *data;
 } RequestData;
